requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-session-timeout==0.1.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
pbr==5.9.0
Pillow==9.1.1
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
tzdata==2021.5
unicorn==1.0.3
urllib3==1.26.11
virtualenv==20.13.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==6.0.0
